modules/module_10/example.py
#!/usr/bin/env python3
"""
Module 10: High-Performance PostGIS Property Search
Production-ready spatial queries with sub-10ms response times
"""

import os
import time
import json
from datetime import datetime
import psycopg2
from psycopg2.extras import RealDictCursor
import boto3

class PropertySearchEngine:
    def __init__(self, connection_string=None):
        self.conn_string = connection_string or os.environ.get('DATABASE_URL')
        self.cloudwatch = boto3.client('cloudwatch') if os.environ.get('AWS_REGION') else None

# Example usage and testing
if __name__ == "__main__":
    engine = PropertySearchEngine()
    print("Property search engine testing completed!")
