requirements.txt
# Core Python web framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database connectivity and ORM
psycopg2-binary==2.9.8
sqlalchemy==2.0.23

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests and API utilities
requests==2.31.0

# Geospatial data processing
geojson==3.0.1
shapely==2.0.2

# Data analysis and manipulation
numpy==1.24.3
pandas==2.0.3

# Date and time utilities
python-dateutil==2.8.2

# JSON processing
jsonschema==4.19.2

# Logging and debugging
structlog==23.2.0

# Testing utilities
pytest==7.4.3
pytest-flask==1.3.0

# Development utilities
black==23.11.0
flake8==6.1.0

# Security utilities
cryptography==41.0.7

# AWS SDK for cloud deployment
boto3==1.29.7
botocore==1.32.7

# Performance monitoring
psutil==5.9.6
